style: Refactor slightly for readability
