+++ src/App.js
@@ -1,20 +1,9 @@
 import * as React from 'react'
 import './vars.css'
 import styled from 'styled-components'
-import { useForm } from "react-hook-form";
+import ThemeToggler from './ThemeToggle'
+import { SUPER_HERO_OPTIONS } from './constants'
 
-// todo: possibly use an api for all heroes
-// todo: at least move to another file
-const SUPER_HERO_OPTIONS = [
-  {
-    id: 'iron-man',
-    headline: 'Iron Man'
-  },
-  {
-    id: 'captain-america',
-    headline: 'Captain America'
-  }
-]
 
 // setting up app state and dispatcher context
 const AppStateContext = React.createContext()
@@ -58,7 +47,7 @@ function useAppState() {
   return context
 }
 
-function useAppDispatch() {
+export function useAppDispatch() {
   const context = React.useContext(AppDispatchContext)
   if (!context) {
     throw new Error('useAppDispatch must be used within the AppProvider')
@@ -87,30 +76,6 @@ const PrimaryText = styled.h1({
   textAlign: 'center',
 })
 
-function ThemeToggler() {
-  const dispatch = useAppDispatch()
-  const { register, handleSubmit } = useForm();
-
-  const onSubmit = data => dispatch({
-    type: 'UPDATE_SUPER_HERO',
-    id: data.theme
-  })
-
-  // should not update each change of pick
-  return (
-    <form onSubmit={handleSubmit(onSubmit)}>
-      <label for="theme">Super hero theme</label>
-      <select id="theme" {...register("theme")}>
-        {SUPER_HERO_OPTIONS.map(({ id, headline }) => {
-          return (
-            <option key={id} value={id}>{headline}</option>
-          )
-        })}
-      </select>
-      <input type="submit" />
-    </form>
-  )
-}
 
 function SuperHeroDisplayText() {
   const state = useAppState()
