+++ src/constants.js
@@ -1,0 +1,16 @@
+// todo: possibly use an api for all heroes
+// todo: at least move to another file
+const SUPER_HERO_OPTIONS = [
+  {
+    id: 'iron-man',
+    headline: 'Iron Man'
+  },
+  {
+    id: 'captain-america',
+    headline: 'Captain America'
+  }
+]
+
+export {
+  SUPER_HERO_OPTIONS
+}
